fix php7 type conflicts - bcs php7 now has native Integer and String and Float ...
