+++ eZ/Publish/Core/Persistence/Legacy/Content/FieldValue/Converter/AuthorConverter.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * File containing the Mail converter
+ * File containing the Author converter
  *
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
@@ -14,17 +14,16 @@ use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldValue;
 use eZ\Publish\SPI\Persistence\Content\FieldValue;
 use eZ\Publish\SPI\Persistence\Content\Type\FieldDefinition;
 use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldDefinition;
+use DOMDocument;
 
-class EmailAddress implements Converter
+class AuthorConverter implements Converter
 {
-    const VALIDATOR_IDENTIFIER = "EmailAddressValidator";
-
     /**
      * Factory for current class
      *
      * @note Class should instead be configured as service if it gains dependencies.
      *
-     * @return \eZ\Publish\Core\Persistence\Legacy\Content\FieldValue\Converter\EmailAddress
+     * @return Author
      */
     public static function create()
     {
@@ -39,8 +38,7 @@ class EmailAddress implements Converter
      */
     public function toStorageValue( FieldValue $value, StorageFieldValue $storageFieldValue )
     {
-        $storageFieldValue->dataText      = $value->data;
-        $storageFieldValue->sortKeyString = $value->sortKey;
+        $storageFieldValue->dataText = $this->generateXmlString( $value->data );
     }
 
     /**
@@ -51,8 +49,7 @@ class EmailAddress implements Converter
      */
     public function toFieldValue( StorageFieldValue $value, FieldValue $fieldValue )
     {
-        $fieldValue->data    = $value->dataText;
-        $fieldValue->sortKey = $value->sortKeyString;
+        $fieldValue->data = $this->restoreValueFromXmlString( $value->dataText );
     }
 
     /**
@@ -63,7 +60,7 @@ class EmailAddress implements Converter
      */
     public function toStorageFieldDefinition( FieldDefinition $fieldDef, StorageFieldDefinition $storageDef )
     {
-        $storageDef->dataText1 = $fieldDef->defaultValue->data;
+        // Nothing to store
     }
 
     /**
@@ -74,10 +71,7 @@ class EmailAddress implements Converter
      */
     public function toFieldDefinition( StorageFieldDefinition $storageDef, FieldDefinition $fieldDef )
     {
-        $validatorConstraints = array( self::VALIDATOR_IDENTIFIER => array() );
-        $fieldDef->fieldTypeConstraints->validators = $validatorConstraints;
-        $fieldDef->defaultValue->data = isset( $storageDef->dataText1 ) ? $storageDef->dataText1 : '';
-        $fieldDef->defaultValue->sortKey = $fieldDef->defaultValue->data;
+        $fieldDef->defaultValue->data = array();
     }
 
     /**
@@ -91,6 +85,63 @@ class EmailAddress implements Converter
      */
     public function getIndexColumn()
     {
-        return 'sort_key_string';
+        return false;
+    }
+
+    /**
+     * Generates XML string from $authorValue to be stored in storage engine
+     *
+     * @param array $authorValue
+     *
+     * @return string The generated XML string
+     */
+    private function generateXmlString( array $authorValue )
+    {
+        $doc = new DOMDocument( '1.0', 'utf-8' );
+
+        $root = $doc->createElement( 'ezauthor' );
+        $doc->appendChild( $root );
+
+        $authors = $doc->createElement( 'authors' );
+        $root->appendChild( $authors );
+
+        foreach ( $authorValue as $author )
+        {
+            $authorNode = $doc->createElement( 'author' );
+            $authorNode->setAttribute( 'id', $author["id"] );
+            $authorNode->setAttribute( 'name', $author["name"] );
+            $authorNode->setAttribute( 'email', $author["email"] );
+            $authors->appendChild( $authorNode );
+            unset( $authorNode );
+        }
+
+        return $doc->saveXML();
+    }
+
+    /**
+     * Restores an author Value object from $xmlString
+     *
+     * @param string $xmlString XML String stored in storage engine
+     *
+     * @return \eZ\Publish\Core\FieldType\Author\Value
+     */
+    private function restoreValueFromXmlString( $xmlString )
+    {
+        $dom = new DOMDocument( '1.0', 'utf-8' );
+        $authors = array();
+
+        if ( $dom->loadXML( $xmlString ) === true )
+        {
+            foreach ( $dom->getElementsByTagName( 'author' ) as $author )
+            {
+                $authors[] = array(
+                    'id' => $author->getAttribute( 'id' ),
+                    'name' => $author->getAttribute( 'name' ),
+                    'email' => $author->getAttribute( 'email' )
+                );
+            }
+        }
+
+        return $authors;
     }
 }
