+++ eZ/Publish/Core/Persistence/Legacy/Content/FieldValue/Converter/CountryConverter.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * File containing the MapLocation converter
+ * File containing the Country converter
  *
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
@@ -11,18 +11,19 @@ namespace eZ\Publish\Core\Persistence\Legacy\Content\FieldValue\Converter;
 
 use eZ\Publish\Core\Persistence\Legacy\Content\FieldValue\Converter;
 use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldValue;
+use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldDefinition;
 use eZ\Publish\SPI\Persistence\Content\FieldValue;
 use eZ\Publish\SPI\Persistence\Content\Type\FieldDefinition;
-use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldDefinition;
+use eZ\Publish\Core\FieldType\FieldSettings;
 
-class MapLocation implements Converter
+class CountryConverter implements Converter
 {
     /**
      * Factory for current class
      *
      * @note Class should instead be configured as service if it gains dependencies.
      *
-     * @return MapLocation
+     * @return Country
      */
     public static function create()
     {
@@ -37,9 +38,8 @@ class MapLocation implements Converter
      */
     public function toStorageValue( FieldValue $value, StorageFieldValue $storageFieldValue )
     {
-        $storageFieldValue->dataInt = isset( $value->externalData['address'] ) ? 1 : 0;
-        $storageFieldValue->dataText = '';
-        $storageFieldValue->sortKeyString = (string)$value->sortKey;
+        $storageFieldValue->dataText = empty( $value->data ) ? "" : implode( ",", $value->data );
+        $storageFieldValue->sortKeyString = $value->sortKey;
     }
 
     /**
@@ -50,6 +50,8 @@ class MapLocation implements Converter
      */
     public function toFieldValue( StorageFieldValue $value, FieldValue $fieldValue )
     {
+        $fieldValue->data = empty( $value->dataText ) ? null : explode( ",", $value->dataText );
+        $fieldValue->sortKey = $value->sortKeyString;
     }
 
     /**
@@ -60,6 +62,14 @@ class MapLocation implements Converter
      */
     public function toStorageFieldDefinition( FieldDefinition $fieldDef, StorageFieldDefinition $storageDef )
     {
+        if ( isset( $fieldDef->fieldTypeConstraints->fieldSettings["isMultiple"] ) )
+        {
+            $storageDef->dataInt1 = (int)$fieldDef->fieldTypeConstraints->fieldSettings["isMultiple"];
+        }
+
+        $storageDef->dataText5 = $fieldDef->defaultValue->data === null
+            ? ""
+            : implode( ",", $fieldDef->defaultValue->data );
     }
 
     /**
@@ -70,6 +80,18 @@ class MapLocation implements Converter
      */
     public function toFieldDefinition( StorageFieldDefinition $storageDef, FieldDefinition $fieldDef )
     {
+        $fieldDef->fieldTypeConstraints->fieldSettings = new FieldSettings(
+            array(
+                "isMultiple" => !empty( $storageDef->dataInt1 ) ? (bool)$storageDef->dataInt1 : false
+            )
+        );
+
+        $fieldDef->defaultValue->data = empty( $storageDef->dataText5 )
+            ? null
+            : explode( ",", $storageDef->dataText5 );
+        // TODO This will contain comma separated country codes, which is correct for value but not for sort key.
+        // Sort key should contain comma separated lowercased country names.
+        $fieldDef->defaultValue->sortKey = $storageDef->dataText5;
     }
 
     /**
@@ -83,7 +105,6 @@ class MapLocation implements Converter
      */
     public function getIndexColumn()
     {
-        return 'sort_key_string';
+        return "sort_key_string";
     }
-
 }
