+++ eZ/Publish/Core/Persistence/Legacy/Content/FieldValue/Converter/IntegerConverter.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * File containing the Checkbox converter
+ * File containing the Integer converter
  *
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
@@ -15,14 +15,19 @@ use eZ\Publish\SPI\Persistence\Content\FieldValue;
 use eZ\Publish\SPI\Persistence\Content\Type\FieldDefinition;
 use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldDefinition;
 
-class Checkbox implements Converter
+class IntegerConverter implements Converter
 {
+    const FLOAT_VALIDATOR_IDENTIFIER = "IntegerValueValidator";
+
+    const HAS_MIN_VALUE = 1;
+    const HAS_MAX_VALUE = 2;
+
     /**
      * Factory for current class
      *
      * @note Class should instead be configured as service if it gains dependencies.
      *
-     * @return Checkbox
+     * @return Integer
      */
     public static function create()
     {
@@ -37,8 +42,8 @@ class Checkbox implements Converter
      */
     public function toStorageValue( FieldValue $value, StorageFieldValue $storageFieldValue )
     {
-        $storageFieldValue->dataInt    = (int)$value->data;
-        $storageFieldValue->sortKeyInt = (int)$value->data;
+        $storageFieldValue->dataInt = $value->data;
+        $storageFieldValue->sortKeyInt = (int)$value->sortKey;
     }
 
     /**
@@ -49,8 +54,8 @@ class Checkbox implements Converter
      */
     public function toFieldValue( StorageFieldValue $value, FieldValue $fieldValue )
     {
-        $fieldValue->data    = (bool)$value->dataInt;
-        $fieldValue->sortKey = $value->dataInt;
+        $fieldValue->data = $value->dataInt;
+        $fieldValue->sortKey = $value->sortKeyInt;
     }
 
     /**
@@ -61,7 +66,19 @@ class Checkbox implements Converter
      */
     public function toStorageFieldDefinition( FieldDefinition $fieldDef, StorageFieldDefinition $storageDef )
     {
-        $storageDef->dataInt3 = (int)$fieldDef->defaultValue->data;
+        if ( isset( $fieldDef->fieldTypeConstraints->validators[self::FLOAT_VALIDATOR_IDENTIFIER]['minIntegerValue'] ) )
+        {
+            $storageDef->dataInt1 = $fieldDef->fieldTypeConstraints->validators[self::FLOAT_VALIDATOR_IDENTIFIER]['minIntegerValue'];
+        }
+
+        if ( isset( $fieldDef->fieldTypeConstraints->validators[self::FLOAT_VALIDATOR_IDENTIFIER]['maxIntegerValue'] ) )
+        {
+            $storageDef->dataInt2 = $fieldDef->fieldTypeConstraints->validators[self::FLOAT_VALIDATOR_IDENTIFIER]['maxIntegerValue'];
+        }
+
+        // Defining dataInt4 which holds the validator state (min value/max value/minMax value)
+        $storageDef->dataInt4 = $this->getStorageDefValidatorState( $storageDef->dataInt1, $storageDef->dataInt2 );
+        $storageDef->dataInt3 = $fieldDef->defaultValue->data;
     }
 
     /**
@@ -72,7 +89,19 @@ class Checkbox implements Converter
      */
     public function toFieldDefinition( StorageFieldDefinition $storageDef, FieldDefinition $fieldDef )
     {
-        $fieldDef->defaultValue->data = !empty( $storageDef->dataInt3 ) ? (bool)$storageDef->dataInt3 : false;
+        $validatorParameters = array( 'minIntegerValue' => false, 'maxIntegerValue' => false );
+        if ( $storageDef->dataInt4 & self::HAS_MIN_VALUE )
+        {
+            $validatorParameters['minIntegerValue'] = $storageDef->dataInt1;
+        }
+
+        if ( $storageDef->dataInt4 & self::HAS_MAX_VALUE )
+        {
+            $validatorParameters['maxIntegerValue'] = $storageDef->dataInt2;
+        }
+        $fieldDef->fieldTypeConstraints->validators[self::FLOAT_VALIDATOR_IDENTIFIER] = $validatorParameters;
+        $fieldDef->defaultValue->data = $storageDef->dataInt3;
+        $fieldDef->defaultValue->sortKey = ( $storageDef->dataInt3 === null ? 0 : $storageDef->dataInt3 );
     }
 
     /**
@@ -89,4 +118,27 @@ class Checkbox implements Converter
         return 'sort_key_int';
     }
 
+    /**
+     * Returns validator state for storage definition.
+     * Validator state is a bitfield value composed of:
+     *   - {@link self::HAS_MAX_VALUE}
+     *   - {@link self::HAS_MIN_VALUE}
+     *
+     * @param int|null $minValue Minimum int value, or null if not set
+     * @param int|null $maxValue Maximum int value, or null if not set
+     *
+     * @return int
+     */
+    private function getStorageDefValidatorState( $minValue, $maxValue )
+    {
+        $state = 0;
+
+        if ( $minValue !== null )
+            $state |= self::HAS_MIN_VALUE;
+
+        if ( $maxValue !== null )
+            $state |= self::HAS_MAX_VALUE;
+
+        return $state;
+    }
 }
