+++ eZ/Publish/Core/Persistence/Legacy/Content/FieldValue/Converter/MapLocationConverter.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * File containing the Author converter
+ * File containing the MapLocation converter
  *
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
@@ -14,16 +14,15 @@ use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldValue;
 use eZ\Publish\SPI\Persistence\Content\FieldValue;
 use eZ\Publish\SPI\Persistence\Content\Type\FieldDefinition;
 use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldDefinition;
-use DOMDocument;
 
-class Author implements Converter
+class MapLocationConverter implements Converter
 {
     /**
      * Factory for current class
      *
      * @note Class should instead be configured as service if it gains dependencies.
      *
-     * @return Author
+     * @return MapLocation
      */
     public static function create()
     {
@@ -38,7 +37,9 @@ class Author implements Converter
      */
     public function toStorageValue( FieldValue $value, StorageFieldValue $storageFieldValue )
     {
-        $storageFieldValue->dataText = $this->generateXmlString( $value->data );
+        $storageFieldValue->dataInt = isset( $value->externalData['address'] ) ? 1 : 0;
+        $storageFieldValue->dataText = '';
+        $storageFieldValue->sortKeyString = (string)$value->sortKey;
     }
 
     /**
@@ -49,7 +50,6 @@ class Author implements Converter
      */
     public function toFieldValue( StorageFieldValue $value, FieldValue $fieldValue )
     {
-        $fieldValue->data = $this->restoreValueFromXmlString( $value->dataText );
     }
 
     /**
@@ -60,7 +60,6 @@ class Author implements Converter
      */
     public function toStorageFieldDefinition( FieldDefinition $fieldDef, StorageFieldDefinition $storageDef )
     {
-        // Nothing to store
     }
 
     /**
@@ -71,7 +70,6 @@ class Author implements Converter
      */
     public function toFieldDefinition( StorageFieldDefinition $storageDef, FieldDefinition $fieldDef )
     {
-        $fieldDef->defaultValue->data = array();
     }
 
     /**
@@ -85,63 +83,7 @@ class Author implements Converter
      */
     public function getIndexColumn()
     {
-        return false;
+        return 'sort_key_string';
     }
 
-    /**
-     * Generates XML string from $authorValue to be stored in storage engine
-     *
-     * @param array $authorValue
-     *
-     * @return string The generated XML string
-     */
-    private function generateXmlString( array $authorValue )
-    {
-        $doc = new DOMDocument( '1.0', 'utf-8' );
-
-        $root = $doc->createElement( 'ezauthor' );
-        $doc->appendChild( $root );
-
-        $authors = $doc->createElement( 'authors' );
-        $root->appendChild( $authors );
-
-        foreach ( $authorValue as $author )
-        {
-            $authorNode = $doc->createElement( 'author' );
-            $authorNode->setAttribute( 'id', $author["id"] );
-            $authorNode->setAttribute( 'name', $author["name"] );
-            $authorNode->setAttribute( 'email', $author["email"] );
-            $authors->appendChild( $authorNode );
-            unset( $authorNode );
-        }
-
-        return $doc->saveXML();
-    }
-
-    /**
-     * Restores an author Value object from $xmlString
-     *
-     * @param string $xmlString XML String stored in storage engine
-     *
-     * @return \eZ\Publish\Core\FieldType\Author\Value
-     */
-    private function restoreValueFromXmlString( $xmlString )
-    {
-        $dom = new DOMDocument( '1.0', 'utf-8' );
-        $authors = array();
-
-        if ( $dom->loadXML( $xmlString ) === true )
-        {
-            foreach ( $dom->getElementsByTagName( 'author' ) as $author )
-            {
-                $authors[] = array(
-                    'id' => $author->getAttribute( 'id' ),
-                    'name' => $author->getAttribute( 'name' ),
-                    'email' => $author->getAttribute( 'email' )
-                );
-            }
-        }
-
-        return $authors;
-    }
 }
