+++ eZ/Publish/Core/Persistence/Legacy/Content/FieldValue/Converter/SelectionConverter.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * File containing the Date field value converter class
+ * File containing the Selection converter
  *
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
@@ -9,26 +9,22 @@
 
 namespace eZ\Publish\Core\Persistence\Legacy\Content\FieldValue\Converter;
 
+use eZ\Publish\Core\FieldType\FieldSettings;
 use eZ\Publish\Core\Persistence\Legacy\Content\FieldValue\Converter;
 use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldValue;
+use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldDefinition;
 use eZ\Publish\SPI\Persistence\Content\FieldValue;
 use eZ\Publish\SPI\Persistence\Content\Type\FieldDefinition;
-use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldDefinition;
-use eZ\Publish\Core\FieldType\Date\Type as DateType;
-use eZ\Publish\Core\FieldType\FieldSettings;
-use DateTime;
+use DOMDocument;
 
-/**
- * Date field value converter class
- */
-class Date implements Converter
+class SelectionConverter implements Converter
 {
     /**
      * Factory for current class
      *
      * @note Class should instead be configured as service if it gains dependencies.
      *
-     * @return \eZ\Publish\Core\Persistence\Legacy\Content\FieldValue\Converter\Date
+     * @return Selection
      */
     public static function create()
     {
@@ -43,8 +39,7 @@ class Date implements Converter
      */
     public function toStorageValue( FieldValue $value, StorageFieldValue $storageFieldValue )
     {
-        $storageFieldValue->dataInt = ( $value->data !== null ? $value->data["timestamp"] : null );
-        $storageFieldValue->sortKeyInt = (int)$value->sortKey;
+        $storageFieldValue->sortKeyString = $storageFieldValue->dataText = $value->sortKey;
     }
 
     /**
@@ -55,16 +50,18 @@ class Date implements Converter
      */
     public function toFieldValue( StorageFieldValue $value, FieldValue $fieldValue )
     {
-        if ( $value->dataInt === null || $value->dataInt == 0 )
+        if ( $value->dataText !== '' )
         {
-            return;
+            $fieldValue->data = array_map(
+                'intval',
+                explode( '-', $value->dataText )
+            );
         }
-
-        $fieldValue->data = array(
-            "timestamp" => $value->dataInt,
-            "rfc850" => null,
-        );
-        $fieldValue->sortKey = $value->sortKeyInt;
+        else
+        {
+            $fieldValue->data = array();
+        }
+        $fieldValue->sortKey = $value->sortKeyString;
     }
 
     /**
@@ -75,7 +72,30 @@ class Date implements Converter
      */
     public function toStorageFieldDefinition( FieldDefinition $fieldDef, StorageFieldDefinition $storageDef )
     {
-        $storageDef->dataInt1 = $fieldDef->fieldTypeConstraints->fieldSettings["defaultType"];
+        $fieldSettings = $fieldDef->fieldTypeConstraints->fieldSettings;
+
+        if ( isset( $fieldSettings["isMultiple"] ) )
+            $storageDef->dataInt1 = (int)$fieldSettings["isMultiple"];
+
+        if ( !empty( $fieldSettings["options"] ) )
+        {
+            $xml = new DOMDocument( "1.0", "utf-8" );
+            $xml->appendChild(
+                $selection = $xml->createElement( "ezselection" )
+            );
+            $selection->appendChild(
+                $options = $xml->createElement( "options" )
+            );
+            foreach ( $fieldSettings["options"] as $id => $name )
+            {
+                $options->appendChild(
+                    $option = $xml->createElement( "option" )
+                );
+                $option->setAttribute( "id", $id );
+                $option->setAttribute( "name", $name );
+            }
+            $storageDef->dataText5 = $xml->saveXML();
+        }
     }
 
     /**
@@ -86,28 +106,28 @@ class Date implements Converter
      */
     public function toFieldDefinition( StorageFieldDefinition $storageDef, FieldDefinition $fieldDef )
     {
+        $options = array();
+        $simpleXml = simplexml_load_string( $storageDef->dataText5 );
+
+        if ( $simpleXml !== false )
+        {
+            foreach ( $simpleXml->options->option as $option )
+            {
+                $options[(int)$option["id"]] = (string)$option["name"];
+            }
+        }
+
         $fieldDef->fieldTypeConstraints->fieldSettings = new FieldSettings(
             array(
-                "defaultType" => $storageDef->dataInt1
+                "isMultiple" => !empty( $storageDef->dataInt1 ) ? (bool)$storageDef->dataInt1 : false,
+                "options" => $options,
             )
         );
 
-        // Building default value
-        switch ( $fieldDef->fieldTypeConstraints->fieldSettings["defaultType"] )
-        {
-            case DateType::DEFAULT_CURRENT_DATE:
-                $dateTime = new DateTime();
-                $dateTime->setTime( 0, 0, 0 );
-                $data = array(
-                    "timestamp" => $dateTime->getTimestamp(),
-                    "rfc850" => null,
-                );
-                break;
-            default:
-                $data = null;
-        }
-
-        $fieldDef->defaultValue->data = $data;
+        // @todo: Can Selection store a default value in the DB?
+        $fieldDef->defaultValue = new FieldValue();
+        $fieldDef->defaultValue->data = array();
+        $fieldDef->defaultValue->sortKey = "";
     }
 
     /**
@@ -121,6 +141,7 @@ class Date implements Converter
      */
     public function getIndexColumn()
     {
-        return "sort_key_int";
+        return "sort_key_string";
     }
+
 }
