+++ eZ/Publish/Core/Persistence/Legacy/Content/FieldValue/Converter/TextLineConverter.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * File containing the Url converter
+ * File containing the TextLine converter
  *
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
@@ -15,14 +15,16 @@ use eZ\Publish\SPI\Persistence\Content\FieldValue;
 use eZ\Publish\SPI\Persistence\Content\Type\FieldDefinition;
 use eZ\Publish\Core\Persistence\Legacy\Content\StorageFieldDefinition;
 
-class Url implements Converter
+class TextLineConverter implements Converter
 {
+    const STRING_LENGTH_VALIDATOR_IDENTIFIER = "StringLengthValidator";
+
     /**
      * Factory for current class
      *
      * @note Class should instead be configured as service if it gains dependencies.
      *
-     * @return Url
+     * @return TextLine
      */
     public static function create()
     {
@@ -37,12 +39,8 @@ class Url implements Converter
      */
     public function toStorageValue( FieldValue $value, StorageFieldValue $storageFieldValue )
     {
-        $storageFieldValue->dataText = isset( $value->data['text'] )
-            ? $value->data['text']
-            : null;
-        $storageFieldValue->dataInt = isset( $value->data['urlId'] )
-            ? $value->data['urlId']
-            : null;
+        $storageFieldValue->dataText = $value->data;
+        $storageFieldValue->sortKeyString = $value->sortKey;
     }
 
     /**
@@ -53,11 +51,8 @@ class Url implements Converter
      */
     public function toFieldValue( StorageFieldValue $value, FieldValue $fieldValue )
     {
-        $fieldValue->data = array(
-            "urlId" => $value->dataInt,
-            'text' => $value->dataText,
-        );
-        $fieldValue->sortKey = false;
+        $fieldValue->data = $value->dataText;
+        $fieldValue->sortKey = $value->sortKeyString;
     }
 
     /**
@@ -68,6 +63,25 @@ class Url implements Converter
      */
     public function toStorageFieldDefinition( FieldDefinition $fieldDef, StorageFieldDefinition $storageDef )
     {
+        if ( isset( $fieldDef->fieldTypeConstraints->validators[self::STRING_LENGTH_VALIDATOR_IDENTIFIER]['maxStringLength'] ) )
+        {
+            $storageDef->dataInt1 = $fieldDef->fieldTypeConstraints->validators[self::STRING_LENGTH_VALIDATOR_IDENTIFIER]['maxStringLength'];
+        }
+        else
+        {
+            $storageDef->dataInt1 = 0;
+        }
+
+        if ( isset( $fieldDef->fieldTypeConstraints->validators[self::STRING_LENGTH_VALIDATOR_IDENTIFIER]['minStringLength'] ) )
+        {
+            $storageDef->dataInt2 = $fieldDef->fieldTypeConstraints->validators[self::STRING_LENGTH_VALIDATOR_IDENTIFIER]['minStringLength'];
+        }
+        else
+        {
+            $storageDef->dataInt2 = 0;
+        }
+
+        $storageDef->dataText1 = $fieldDef->defaultValue->data;
     }
 
     /**
@@ -78,9 +92,26 @@ class Url implements Converter
      */
     public function toFieldDefinition( StorageFieldDefinition $storageDef, FieldDefinition $fieldDef )
     {
-        // @todo: Is it possible to store a default value in the DB?
-        $fieldDef->defaultValue = new FieldValue();
-        $fieldDef->defaultValue->data = array( 'text' => null );
+        $validatorConstraints = array();
+
+        if ( isset( $storageDef->dataInt1 ) )
+        {
+            $validatorConstraints[self::STRING_LENGTH_VALIDATOR_IDENTIFIER]["maxStringLength"] =
+                $storageDef->dataInt1 != 0 ?
+                    (int)$storageDef->dataInt1 :
+                    false;
+        }
+        if ( isset( $storageDef->dataInt2 ) )
+        {
+            $validatorConstraints[self::STRING_LENGTH_VALIDATOR_IDENTIFIER]["minStringLength"] =
+                $storageDef->dataInt2 != 0 ?
+                    (int)$storageDef->dataInt2 :
+                    false;
+        }
+
+        $fieldDef->fieldTypeConstraints->validators = $validatorConstraints;
+        $fieldDef->defaultValue->data = $storageDef->dataText1 ?: null;
+        $fieldDef->defaultValue->sortKey = $storageDef->dataText1 ?: "";
     }
 
     /**
@@ -90,10 +121,10 @@ class Url implements Converter
      * "sort_key_int" or "sort_key_string". This column is then used for
      * filtering and sorting for this type.
      *
-     * @return false
+     * @return string
      */
     public function getIndexColumn()
     {
-        return false;
+        return 'sort_key_string';
     }
 }
